+++ changelog.md
@@ -2,7 +2,7 @@
 
 ## Upcoming
 * Shields plus data pack no longer enables itself if the mod is not present
-* Moved Season plant growth modifier to ITR
+* Moved Season plant growth modifier and livestock to ITR
 
 ## 1.11.6
 * Fixed possible startup crash